randomly dodge reduction
